+++ lib/shakapacker/version_checker.rb
@@ -126,7 +126,7 @@ module Shakapacker
           end
 
           def relative_path?
-            raw.match(%r{(\.\.|\Afile:///)}).present?
+            raw.match(%r{(\.\.|\Afile:)}).present?
           end
 
           def git_url?