+++ package/rules/esbuild.js
@@ -1,9 +1,9 @@
 const { loaderMatches } = require("../utils/helpers")
 const { getEsbuildLoaderConfig } = require("../esbuild")
-const { webpack_loader: webpackLoader } = require("../config")
+const { javascript_transpiler: javascriptTranspiler } = require("../config")
 const jscommon = require("./jscommon")
 
-module.exports = loaderMatches(webpackLoader, "esbuild", () => ({
+module.exports = loaderMatches(javascriptTranspiler, "esbuild", () => ({
   test: /\.(ts|tsx|js|jsx|mjs|coffee)?(\.erb)?$/,
   ...jscommon,
   use: ({ resource }) => getEsbuildLoaderConfig(resource)