+++ package/rules/swc.js
@@ -1,9 +1,9 @@
 const { loaderMatches } = require("../utils/helpers")
 const { getSwcLoaderConfig } = require("../swc")
-const { webpack_loader: webpackLoader } = require("../config")
+const { javascript_transpiler: javascriptTranspiler } = require("../config")
 const jscommon = require("./jscommon")
 
-module.exports = loaderMatches(webpackLoader, "swc", () => ({
+module.exports = loaderMatches(javascriptTranspiler, "swc", () => ({
   test: /\.(ts|tsx|js|jsx|mjs|coffee)?(\.erb)?$/,
   ...jscommon,
   use: ({ resource }) => getSwcLoaderConfig(resource)